ECS model for items
